Part the Third of the War of the waRnINGs:
Using Clang 3.1 with the -Wall flag in debug mode, the following libraries
should be free of warnings:

basic ~ Mpacella 13:48, 1 October 2012 (EDT)
numeric ~ Mpacella 14:38, 1 October 2012 (EDT)
utility ~ Mpacella 13:35, 11 October 2012 (EDT)
core
    core.1 (no warnings)
    core.2 ~ Labonte 14:46, 2 October 2012 (EDT)
    core.3 ~ Labonte 14:19, 5 October 2012 (EDT)
    core.4 ~ Labonte 16:01, 5 October 2012 (EDT)
    core.5 ~ Labonte 15:10, 5 October 2012 (EDT)
protocols
    protocols.1 ~ Labonte 16:57, 16 October 2012 (EDT)
    protocols.2
        protocols.2a ~ Labonte 17:48, 17 October 2012 (EDT)
        protocols.2b (no warnings)
    protocols.3 ~ Labonte 18:17, 18 October 2012 (EDT)
    protocols.4
        protocols.4a ~ Labonte 18:45, 22 October 2012 (EDT)
        protocols.4b ~ Labonte 14:33, 8 December 2012 (EST)
        protocols.4c ~ Labonte 16:31, 8 December 2012 (EST)
        protocols.4d (no warnings)
        protocols.4e (no warnings)
        protocols.4f ~ Labonte 15:36, 9 December 2012 (EST)
        protocols.4g ~ Labonte 15:41, 9 December 2012 (EST)
        protocols.4h ~ Labonte 16:01, 9 December 2012 (EST)
    protocols.5
        protocols.5a ~ weitzner 16:13, 9 December 2012 (EST)
        protocols.5b ~ Labonte 20:05, 9 December 2012 (EST)
        protocols.5c ~ weitzner 16:03, 9 December 2012 (EST)
        protocols.5d ~ weitzner 19:34, 9 December 2012 (EST)
        protocols.5e (32 remaining warnings all in included BOOST headers) ~ weitzner 19:41, 9 December 2012 (EST)
        protocols.5f ~ weitzner 16:01, 9 December 2012 (EST)
    protocols.6 (three #pragma warnings remain) weitzner 15:53, 9 December 2012 (EST)
    protocols.7 (no warnings)
devel ~ weitzner 20:06, 9 December 2012 (EST)

Only apps remain. Then we need to look at warnings in release mode and with GCC....

All unit tests passed; no integration test changes.

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@52578 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: 396ee0c5ed6af86055572a59dc0cacc71714d2a2

+++ rosetta_source/src/protocols/medal/util.cc
@@ -92,12 +92,12 @@ void cutpoint_probabilities(const unsigned num_residues, const core::kinematics:
 
   // List of cutpoints sorted by position
   std::set<unsigned> cutpoints;
-  for (unsigned i = 1; i <= tree.num_cutpoint(); ++i) {
+  for (int i = 1; i <= tree.num_cutpoint(); ++i) {
     unsigned cutpoint = tree.cutpoint(i);
     if (cutpoint == num_residues)  // cutpoint at end of chain
       continue;
 
-    cutpoints.insert(cutpoint);
+    cutpoints.insert((unsigned)cutpoint);
   }
 
   p->clear();
@@ -114,7 +114,7 @@ void cutpoint_probabilities(const unsigned num_residues, const core::kinematics:
 void invalidate_residues_spanning_cuts(const core::kinematics::FoldTree& tree,
                                        const core::Size fragment_len,
                                        Probabilities* probs) {
-  for (unsigned i = 1; i <= tree.num_cutpoint(); ++i) {
+  for (int i = 1; i <= tree.num_cutpoint(); ++i) {
     const unsigned cutpoint = tree.cutpoint(i);
     for (unsigned j = (cutpoint - fragment_len + 2); j <= cutpoint; ++j) {
       (*probs)[j] = 0;
